pyproject.toml
[project]
name = "mini-crm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.3.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "tinydb>=4.8.2",
    "uvicorn>=0.35.0",
    "veris-cli>=1.4.2",
    "veris-ai[fastapi,instrument]>=1.10.1",
]
